backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
python-multipart
matplotlib
numpy
